Rename watershed and _watershed modules respectively to _watershed and _watershed_cy
