+++ skimage/segmentation/setup.py
@@ -11,11 +11,11 @@ def configuration(parent_package='', top_path=None):
 
     config = Configuration('segmentation', parent_package, top_path)
 
-    cython(['_watershed.pyx',
+    cython(['_watershed_cy.pyx',
             '_felzenszwalb_cy.pyx',
             '_quickshift_cy.pyx',
             '_slic.pyx'], working_path=base_path)
-    config.add_extension('_watershed', sources=['_watershed.c'],
+    config.add_extension('_watershed_cy', sources=['_watershed_cy.c'],
                          include_dirs=[get_numpy_include_dirs()])
     config.add_extension('_felzenszwalb_cy', sources=['_felzenszwalb_cy.c'],
                          include_dirs=[get_numpy_include_dirs()])
